post update and recursive delete

+++ app/Http/Controllers/AdminUsersController.php
@@ -107,7 +107,11 @@ class AdminUsersController extends Controller
         
         $user = User::findOrFail($id);
 
-        unlink(public_path() . $user->photo->file); //deletes file 
+        if ($user->photo){
+
+           unlink(public_path() . $user->photo->file); //deletes file 
+
+        }
 
         $user->delete();
